+++ src/main/java/com/snowflake/kafka/connector/SnowflakeSinkConnector.java
@@ -63,7 +63,8 @@ public class SnowflakeSinkConnector extends SinkConnector {
   // Using setupComplete to synchronize
   private boolean setupComplete;
 
-  private final IConnectorConfigValidator connectorConfigValidator = new ConnectorConfigValidator();
+  private final ConnectorConfigValidator connectorConfigValidator =
+      new DefaultConnectorConfigValidator();
 
   /** No-Arg constructor. Required by Kafka Connect framework */
   public SnowflakeSinkConnector() {
