+++ src/test/java/com/snowflake/kafka/connector/DefaultConnectorConfigValidatorTest.java
@@ -37,9 +37,10 @@ import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.MethodSource;
 
-public class ConnectorConfigValidatorTest {
+public class DefaultConnectorConfigValidatorTest {
 
-  private final IConnectorConfigValidator connectorConfigValidator = new ConnectorConfigValidator();
+  private final ConnectorConfigValidator connectorConfigValidator =
+      new DefaultConnectorConfigValidator();
 
   // subset of valid community converters
   public static final List<Converter> COMMUNITY_CONVERTER_SUBSET =
