+++ src/test/java/com/snowflake/kafka/connector/internal/TombstoneRecordIngestionIT.java
@@ -1,7 +1,7 @@
 package com.snowflake.kafka.connector.internal;
 
-import static com.snowflake.kafka.connector.ConnectorConfigValidatorTest.COMMUNITY_CONVERTER_SUBSET;
-import static com.snowflake.kafka.connector.ConnectorConfigValidatorTest.CUSTOM_SNOWFLAKE_CONVERTERS;
+import static com.snowflake.kafka.connector.DefaultConnectorConfigValidatorTest.COMMUNITY_CONVERTER_SUBSET;
+import static com.snowflake.kafka.connector.DefaultConnectorConfigValidatorTest.CUSTOM_SNOWFLAKE_CONVERTERS;
 
 import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Sets;
